add console producer/consumer
