+++ solutions/telemetry/certs/.gitignore
@@ -1,0 +1,4 @@
+*
+*/
+!.gitignore
+
