+++ solutions/telemetry/console/consumer/README.md
@@ -1,0 +1,13 @@
+# Kafka Basic Consumer App
+
+First, build the app:
+
+```
+./mvnw package
+```
+
+Next, run the app:
+
+```
+java -jar target/consumer-1.0-SNAPSHOT.jar
+```
