+++ solutions/telemetry/console/producer/README.md
@@ -1,0 +1,13 @@
+# Kafka Basic Producer App
+
+First, build the app:
+
+```
+./mvnw package
+```
+
+Next, run the app:
+
+```
+java -jar target/producer-1.0-SNAPSHOT.jar
+```
