--- telemetry/certs/ca.crt
@@ -1,20 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDLTCCAhWgAwIBAgIJAOuKA5dDFbR3MA0GCSqGSIb3DQEBCwUAMC0xEzARBgNV
-BAoMCmlvLnN0cmltemkxFjAUBgNVBAMMDWNsdXN0ZXItY2EgdjAwHhcNMjEwNDIy
-MTQwNjQ0WhcNMjIwNDIyMTQwNjQ0WjAtMRMwEQYDVQQKDAppby5zdHJpbXppMRYw
-FAYDVQQDDA1jbHVzdGVyLWNhIHYwMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIB
-CgKCAQEAy78n1tRO5cH3yhofD/+R7CWjDFnnuAFc+vryOqd+zuAN0wmih4UInXp9
-iYlTJ9bv3g21dpyYiL5jR/3d1gigpAlL1qlWcRoLe59mjAtW49f7GJqJtGDobVlN
-DftYbuTg6jP9L1BuDa97/BbbNV9sa5KHZtRmZ/Ob1BCU5rarS3e5NabiPXna7+aD
-igtDqOwKUIUw8y5gMOwvVp4KJV66hxzrz8S53xv/QHI3+u/Yb8aYYanVV+VgbZL3
-2OPU7ryKJHU1zxUCJAWdw51WNBKllV2eBW7ocUSmSANoiyFwrr4NEeXvjIlIr0vA
-yQ5Qrl2P/x2vSF4u5pw+jX5Aj01sJQIDAQABo1AwTjAdBgNVHQ4EFgQUF2wZGagZ
-XWspwmZk3sXvwbb3zgYwHwYDVR0jBBgwFoAUF2wZGagZXWspwmZk3sXvwbb3zgYw
-DAYDVR0TBAUwAwEB/zANBgkqhkiG9w0BAQsFAAOCAQEADsoVN3FxTfMm8+yLhreG
-fzvASwKN7Yy0txcFPiHvehEuMaRD8NcyNZFdDSwYPcV1vAZTvGAZCS/j7pImZw9c
-rTuMGH8GM8Nm6/ixUGpBpm2VcZPNAumYhdH2pYuvjbX6ydraaPpfjjXMeAo+UI64
-3114v/nlG07kxyKLYDM5p/akNcW/55FdmI7dzzYC9+jRvI+9qz2zgHGUoXU0qdwu
-LwrtadBdBgAWYJ1RQmfFnXFemA9Nh+bjHqvv3XycLYdKpqQ/GSk6kzmskqg+Tqqw
-IuoAxybpohUjVoakoBtkg7bzumNRigx5/8rlTgXxfXUxekrDHsd0J19OHPKB6Jts
-bQ==
------END CERTIFICATE-----
