--- telemetry/src/main/resources/application.properties
@@ -1,22 +1,0 @@
-
-
-mp.messaging.outgoing.temperatures-produced.connector=smallrye-kafka
-mp.messaging.outgoing.temperatures-produced.topic=temperatures
-# mp.messaging.outgoing.temperatures.key.serializer=org.apache.kafka.common.serialization.IntegerSerializer
-mp.messaging.outgoing.temperatures-produced.value.serializer=org.apache.kafka.common.serialization.IntegerSerializer
-
-
-mp.messaging.incoming.temperatures.connector=smallrye-kafka
-mp.messaging.incoming.temperatures.topic=temperatures
-# mp.messaging.incoming.temperatures.key.deserializer=org.apache.kafka.common.serialization.IntegerDeserializer
-mp.messaging.incoming.temperatures.value.deserializer=org.apache.kafka.common.serialization.IntegerDeserializer
-mp.messaging.incoming.temperatures.auto.offset.reset=earliest
-mp.messaging.incoming.temperatures.group.id=temperatures
-
-
-kafka.bootstrap.servers=my-cluster-kafka-bootstrap-jramirez-kafka.apps.na46-stage2.dev.nextcle.com:443
-kafka.security.protocol=SSL
-kafka.ssl.keystore.location=../certs/keystore.jks
-kafka.ssl.keystore.password=password
-kafka.ssl.truststore.location=../certs/truststore.jks
-kafka.ssl.truststore.password=password
