--- telemetry/src/test/java/org/redhat/telemetry/NativeTelemetryResourceIT.java
@@ -1,9 +1,0 @@
-package org.redhat.telemetry;
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class NativeTelemetryResourceIT extends TelemetryResourceTest {
-
-    // Execute the same tests but in native mode.
-}
